Bugfix on network settings.
Implemented own error checking on copy command using an additional diff
command. This compares the two files that previously has been copied.
If both files are the same the return value equals null, this triggers
an NullReferenceException. The exception is catched and code can
continiue to run.

If copy failed, the return value contains all differences, in that case
i return false so the main function knows to show an error message.